+++ SpaceShooter/Assets/Scripts/Health/IHealable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHealable 
 {
-    public void Heal(int amount);
+    public void Heal(float amount);
 }
